Fix the PyLoop_* macros and work on the proper declaration and definition of module globals.
We avoid symbols conflicts by defining the globals using the INIT() method which should only
be used once per module regardless of the consistency of the source files that make up the module.

Extern declarations are used to allow access to them from arbitrary sources.

This change was necessary in order to support multiple source file modules that development is seeking to support.