__func__ is not a define at any level.
In other runs using this version, this appeared to be working. Recently,
however, this caused compilation failures.

snprintf the dictionary key in response. For now. The override mechanism
as is can be rather taxing for fault.traffic with GIL ensure surrounding
key checks. Future work should look to using a custom hash that doesn't
rely on Python objects.