Refactor libconstruct.
	- Migrate to &..filesystem storage for build directories.
	- Allow for arbitrary variants (work directory key) to decide the build location.
	- Get rid of old formats handling out/&&format/transformed; use variants instead.
	- Simplify induct/construct by moving common parts to libconstruct
	- Use XInclude processing for managing inheritance in context storage files.
	- Normalize management of XML schemas to be similar to &..system.

Fixes the problem with formats in that differing formats are essentially a new build,
not a subdirectory in ./out.

Fragments are likely broken and there are test failures, but there is more work to be
done in order to provide the tests with a fake context set for testing purposes.