Introduce the Factor (target) type.
 - Various regression fixes.
 - Factor type for querying builds and providing target metadata..
 - Provides enclosure for many of the libconstruct APIs
 - Introduction of the 'source' factor type for include directories
 - Allowance of mechanisms to define ignored file extensions.
 - Provide a full index of typed (factor) references for mechanisms to use.

The explicit target type was needed in order to allow simulated/mocked factors
to work properly. The repeated resolution of information relative to an Import,
ModuleType, or module file path made it rather difficult abstract. An actual
type referring these identifiers and definitions allows any part to be
resolved prior to instantiating the factor. This allows virtual/simulated
factors to used.

There are still failing tests, but this is allowed as libconstruct
is still changing.