Preparations for v2 (vector compositions) and eliminate associated bitrot.
- Eliminate domain usage in favor of type caches.
- Eliminate the "formats" field in stored mechanisms in favor of scanning "integrations".
- Simplify options processing given changes to semantics.
- Simplify SystemFactor and use if.fault.io/factors types.
- Relocate Mechanism back into v1 as v2's implementation differs.
- Relocate default source parameters to command constructor implementation.
- Implement Integrand.select for providing (flat) access to factor data.
- Elimiate per-language descriptor parameters.

A general Mechanism implementation is possible, and the original intent
of placing it in core will likely happen in the near future once
the transition is complete. However, the implementation is tightly tied to v1, so
relocate it in preparation for its removal.

Per-language descriptor parameters were intended to provide link constructors with a profile
of the objects being combined: information about the languages/compilers that may have
additional requirements for the linker. In v2, this will likely be handled indirectly
by injecting option productions defined by a language configuration or not at all and
force runtime requirements to be designated by the factor.